refactor: [SNOW-3017492] split reporters
