+++ src/snowflake/cli/_plugins/dcm/reporters/__init__.py
@@ -1,0 +1,18 @@
+# Copyright (c) 2024 Snowflake Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from snowflake.cli._plugins.dcm.reporters.base import Reporter
+from snowflake.cli._plugins.dcm.reporters.refresh import RefreshReporter
+from snowflake.cli._plugins.dcm.reporters.test import TestReporter
+
+__all__ = ["Reporter", "RefreshReporter", "TestReporter"]
