+++ src/snowflake/cli/_plugins/dcm/reporters/refresh.py
@@ -13,233 +13,18 @@
 # limitations under the License.
 import json
 import logging
-from abc import ABC, abstractmethod
 from dataclasses import dataclass, field
 from enum import Enum
-from typing import Any, Dict, Generic, Iterator, List, Optional, Protocol, TypeVar
+from typing import Any, Dict, Iterator, List, Optional, Protocol
 
 from pydantic import BaseModel, Field, ValidationError
 from rich.text import Text
 from snowflake.cli._plugins.dcm import styles
-from snowflake.cli.api.console.console import cli_console
+from snowflake.cli._plugins.dcm.reporters.base import Reporter, cli_console
 from snowflake.cli.api.exceptions import CliError
 from snowflake.cli.api.sanitizers import sanitize_for_terminal
-from snowflake.connector.cursor import SnowflakeCursor
 
 log = logging.getLogger(__name__)
-
-T = TypeVar("T")
-
-
-class Reporter(ABC, Generic[T]):
-    def __init__(self) -> None:
-        self.command_name = ""
-
-    @abstractmethod
-    def extract_data(self, result_json: Dict[str, Any]) -> List[Dict[str, Any]]:
-        """Extract the relevant data from the result JSON."""
-        ...
-
-    @abstractmethod
-    def parse_data(self, data: List[Dict[str, Any]]) -> Iterator[T]:
-        """Parse raw data into domain objects."""
-        ...
-
-    @abstractmethod
-    def print_renderables(self, data: Iterator[T]) -> None:
-        """Print Rich renderables for the parsed data."""
-        ...
-
-    @abstractmethod
-    def _is_success(self) -> bool:
-        """Check if underlying operation passed without errors"""
-        ...
-
-    @abstractmethod
-    def _generate_summary_renderables(self) -> List[Text]:
-        """Generate a list of rich renderables to be printed as success or error message"""
-        ...
-
-    def print_summary(self) -> None:
-        """Print operation summary when the result is successful."""
-        renderables = self._generate_summary_renderables()
-        cli_console.styled_message("\n")
-        for renderable in renderables:
-            cli_console.styled_message(renderable.plain, style=renderable.style)
-        cli_console.styled_message("\n")
-
-    def process(self, cursor: SnowflakeCursor) -> None:
-        """Process cursor data and print results."""
-        row = cursor.fetchone()
-        if not row:
-            cli_console.styled_message("No data.\n")
-            return
-
-        try:
-            result_data = row[0]
-            result_json = (
-                json.loads(result_data) if isinstance(result_data, str) else result_data
-            )
-        except IndexError:
-            log.debug("Unexpected response format: %s", row)
-            raise CliError("Could not process response.")
-        except json.JSONDecodeError as e:
-            log.debug("Could not decode response: %s", e)
-            raise CliError("Could not process response.")
-
-        raw_data = self.extract_data(result_json)
-        parsed_data: Iterator[T] = self.parse_data(raw_data)
-        self.print_renderables(parsed_data)
-        if self._is_success():
-            self.print_summary()
-        else:
-            message = "".join(
-                renderable.plain for renderable in self._generate_summary_renderables()
-            )
-            raise CliError(message)
-
-
-class TestStatus(Enum):
-    __test__ = False  # Prevent pytest collection
-
-    UNKNOWN = "UNKNOWN"
-    PASS = "PASS"
-    FAIL = "FAIL"
-
-
-@dataclass
-class TestRow:
-    __test__ = False  # Prevent pytest collection
-
-    table_name: str = "UNKNOWN"
-    expectation_name: str = "UNKNOWN"
-    status: TestStatus = TestStatus.UNKNOWN
-    expectation_expression: str = ""
-    metric_name: str = ""
-    actual_value: str = ""
-
-    @classmethod
-    def from_dict(cls, data: Dict[str, Any]) -> Optional["TestRow"]:
-        def _get(key):
-            return sanitize_for_terminal(str(data.get(key, "UNKNOWN")))
-
-        if not isinstance(data, dict):
-            log.debug("Unexpected test entry type: %s", type(data))
-            return None
-
-        row = cls(
-            table_name=_get("table_name"),
-            expectation_name=_get("expectation_name"),
-            expectation_expression=_get("expectation_expression"),
-            metric_name=_get("metric_name"),
-            actual_value=_get("value"),
-        )
-
-        expectation_violated = data.get("expectation_violated")
-        if expectation_violated is True:
-            row.status = TestStatus.FAIL
-        elif expectation_violated is False:
-            row.status = TestStatus.PASS
-        else:
-            row.status = TestStatus.UNKNOWN
-        return row
-
-
-class TestReporter(Reporter[TestRow]):
-    __test__ = False  # Prevent pytest collection
-
-    STATUS_WIDTH = 11
-    _DATA_KEY = "expectations"
-
-    @dataclass
-    class Summary:
-        passed: int = 0
-        failed: int = 0
-        unknown: int = 0
-
-        @property
-        def total(self):
-            return self.passed + self.failed + self.unknown
-
-    def __init__(self):
-        super().__init__()
-        self.command_name = "test"
-        self._summary = self.Summary()
-
-    def extract_data(self, result_json: Dict[str, Any]) -> List[Dict[str, Any]]:
-        if not isinstance(result_json, dict):
-            log.debug("Unexpected response type: %s, expected dict", type(result_json))
-            raise CliError("Could not process response.")
-
-        expectations = result_json.get(self._DATA_KEY, list())
-
-        if not isinstance(expectations, list):
-            log.warning(
-                "Unexpected expectations type: %s, expected list",
-                type(expectations),
-            )
-            raise CliError("Could not process response.")
-
-        return expectations
-
-    def parse_data(self, data: List[Dict[str, Any]]) -> Iterator[TestRow]:
-        for row in data:
-            parsed = TestRow.from_dict(row)
-            if parsed is not None:
-                if parsed.status == TestStatus.PASS:
-                    self._summary.passed += 1
-                elif parsed.status == TestStatus.FAIL:
-                    self._summary.failed += 1
-                else:
-                    self._summary.unknown += 1
-                yield parsed
-
-    def print_renderables(self, data: Iterator[TestRow]) -> None:
-        for row in data:
-            if row.status == TestStatus.PASS:
-                status_text = "✓ PASS"
-                style = styles.PASS_STYLE
-            elif row.status == TestStatus.FAIL:
-                status_text = "✗ FAIL"
-                style = styles.FAIL_STYLE
-            else:
-                status_text = "? UNKNOWN"
-                style = styles.STATUS_STYLE
-
-            cli_console.styled_message(
-                status_text.ljust(self.STATUS_WIDTH) + " ",
-                style=style,
-            )
-            cli_console.styled_message(row.table_name, style=styles.DOMAIN_STYLE)
-            cli_console.styled_message(f" ({row.expectation_name})")
-            cli_console.styled_message("\n")
-
-            if row.status == TestStatus.FAIL:
-                cli_console.styled_message(
-                    f"  └─ Expected: {row.expectation_expression}, "
-                    f"Got: {row.actual_value} (Metric: {row.metric_name})\n"
-                )
-
-    def _generate_summary_renderables(self) -> List[Text]:
-        total = self._summary.total
-        if total == 0:
-            return [Text("No expectations found in the project.")]
-
-        result = [
-            (Text(f"{self._summary.passed} passed", styles.PASS_STYLE)),
-            (Text(", ")),
-            (Text(f"{self._summary.failed} failed", styles.FAIL_STYLE)),
-        ]
-        if self._summary.unknown > 0:
-            result.append(Text(", "))
-            result.append(Text(f"{self._summary.unknown} unknown", styles.FAIL_STYLE))
-        result.append(Text(" out of "))
-        result.append(Text(f"{total}", styles.BOLD_STYLE))
-        result.append(Text(" total."))
-        return result
-
-    def _is_success(self) -> bool:
-        return self._summary.failed + self._summary.unknown == 0
 
 
 class RefreshStatistics(BaseModel):
