+++ tests/dcm/test_reporters/test_refresh_reporter.py
@@ -11,63 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import json
-from io import StringIO
-from unittest import mock
-
 import pytest
-from snowflake.cli._plugins.dcm.reporters import (
+from snowflake.cli._plugins.dcm.reporters.refresh import (
     NewFormatExtractor,
     OldFormatExtractor,
     RefreshReporter,
     RefreshRow,
-    TestReporter,
-    TestRow,
-    TestStatus,
 )
-from snowflake.cli.api.exceptions import CliError
 
-
-class FakeCursor:
-    """Fake cursor that returns JSON data like a real Snowflake cursor."""
-
-    def __init__(self, data):
-        self._data = data
-        self._fetched = False
-
-    def fetchone(self):
-        if self._fetched:
-            return None
-        self._fetched = True
-        if self._data is None:
-            return None
-        return (json.dumps(self._data) if isinstance(self._data, dict) else self._data,)
-
-
-def capture_reporter_output(reporter, cursor):
-    """Capture the output from a reporter's process method."""
-    output = StringIO()
-
-    def mock_print(text, style=""):
-        if hasattr(text, "plain"):
-            output.write(text.plain)
-        else:
-            output.write(str(text))
-
-    error_message = ""
-    with mock.patch(
-        "snowflake.cli._plugins.dcm.reporters.cli_console.styled_message",
-        side_effect=mock_print,
-    ):
-        try:
-            reporter.process(cursor)
-        except CliError as e:
-            error_message = e.message
-
-    result = output.getvalue()
-    if error_message:
-        result += f"\n{error_message}\n"
-    return result
+from tests.dcm.test_reporters.utils import FakeCursor, capture_reporter_output
 
 
 class TestRefreshReporter:
@@ -118,7 +70,10 @@ class TestRefreshReporter:
                 },
             ]
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_large_numbers(self, snapshot):
@@ -134,7 +89,10 @@ class TestRefreshReporter:
                 },
             ]
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_statistics_as_dict(self, snapshot):
@@ -146,7 +104,10 @@ class TestRefreshReporter:
                 }
             ]
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_only_insertions(self, snapshot):
@@ -158,7 +119,10 @@ class TestRefreshReporter:
                 }
             ]
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_only_deletions(self, snapshot):
@@ -170,21 +134,33 @@ class TestRefreshReporter:
                 }
             ]
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_empty_cursor(self, snapshot):
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(None))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(None),
+        )
         assert output == snapshot
 
     def test_no_dynamic_tables(self, snapshot):
         data = {"refreshed_tables": []}
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_missing_refreshed_tables_key(self, snapshot):
         data = {"some_other_key": "value"}
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_malformed_statistics_json(self, snapshot):
@@ -196,7 +172,10 @@ class TestRefreshReporter:
                 }
             ]
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_unknown_statistics_format(self, snapshot):
@@ -208,7 +187,10 @@ class TestRefreshReporter:
                 }
             ]
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_null_values_in_statistics(self, snapshot):
@@ -220,7 +202,10 @@ class TestRefreshReporter:
                 }
             ]
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_missing_dt_name(self, snapshot):
@@ -231,7 +216,10 @@ class TestRefreshReporter:
                 }
             ]
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_non_dict_table_entries(self, snapshot):
@@ -241,7 +229,10 @@ class TestRefreshReporter:
                 {"dt_name": "DB.SCHEMA.VALID", "statistics": "No new data"},
             ]
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_string_numbers_in_statistics(self, snapshot):
@@ -253,7 +244,10 @@ class TestRefreshReporter:
                 }
             ]
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_ansi_codes_in_table_name(self, snapshot):
@@ -265,7 +259,10 @@ class TestRefreshReporter:
                 }
             ]
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
 
@@ -309,7 +306,10 @@ class TestNewRefreshReporter:
                 ]
             }
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_single_up_to_date_table(self, snapshot):
@@ -324,7 +324,10 @@ class TestNewRefreshReporter:
                 ]
             }
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_multiple_tables_mixed_status(self, snapshot):
@@ -350,7 +353,10 @@ class TestNewRefreshReporter:
                 ]
             }
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_large_numbers(self, snapshot):
@@ -374,7 +380,10 @@ class TestNewRefreshReporter:
                 ]
             }
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_only_insertions(self, snapshot):
@@ -388,7 +397,10 @@ class TestNewRefreshReporter:
                 ]
             }
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_only_deletions(self, snapshot):
@@ -402,16 +414,25 @@ class TestNewRefreshReporter:
                 ]
             }
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_empty_cursor(self, snapshot):
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(None))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(None),
+        )
         assert output == snapshot
 
     def test_no_dynamic_tables(self, snapshot):
         data = {"dts_refresh_result": {"refreshed_tables": []}}
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_missing_statistics(self, snapshot):
@@ -420,7 +441,10 @@ class TestNewRefreshReporter:
                 "refreshed_tables": [{"table_name": "DB.SCHEMA.NO_STATS"}]
             }
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_missing_table_name(self, snapshot):
@@ -431,7 +455,10 @@ class TestNewRefreshReporter:
                 ]
             }
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_ansi_codes_in_table_name(self, snapshot):
@@ -445,7 +472,10 @@ class TestNewRefreshReporter:
                 ]
             }
         }
-        output = capture_reporter_output(RefreshReporter(), FakeCursor(data))
+        output = capture_reporter_output(
+            RefreshReporter(),
+            FakeCursor(data),
+        )
         assert output == snapshot
 
     def test_selects_correct_extractor_for_new_format(self):
@@ -461,209 +491,3 @@ class TestNewRefreshReporter:
         extractor_cls = reporter._get_extractor_cls(data)  # noqa: SLF001
 
         assert extractor_cls is OldFormatExtractor
-
-
-class TestTestReporter:
-    def test_single_passing_expectation(self, snapshot):
-        data = {
-            "expectations": [
-                {
-                    "table_name": "DB.SCHEMA.CUSTOMERS",
-                    "expectation_name": "row_count_check",
-                    "expectation_violated": False,
-                    "expectation_expression": "> 0",
-                    "metric_name": "row_count",
-                    "value": "1500",
-                }
-            ]
-        }
-        output = capture_reporter_output(TestReporter(), FakeCursor(data))
-        assert output == snapshot
-
-    def test_single_failing_expectation(self, snapshot):
-        data = {
-            "expectations": [
-                {
-                    "table_name": "DB.SCHEMA.ORDERS",
-                    "expectation_name": "null_check",
-                    "expectation_violated": True,
-                    "expectation_expression": "= 0",
-                    "metric_name": "null_count",
-                    "value": "15",
-                }
-            ]
-        }
-        output = capture_reporter_output(TestReporter(), FakeCursor(data))
-        assert output == snapshot
-
-    def test_multiple_expectations_mixed_status(self, snapshot):
-        data = {
-            "expectations": [
-                {
-                    "table_name": "DB.SCHEMA.TABLE_A",
-                    "expectation_name": "unique_check",
-                    "expectation_violated": False,
-                },
-                {
-                    "table_name": "DB.SCHEMA.TABLE_B",
-                    "expectation_name": "not_null",
-                    "expectation_violated": True,
-                    "expectation_expression": "= 0",
-                    "metric_name": "null_count",
-                    "value": "42",
-                },
-                {
-                    "table_name": "DB.SCHEMA.TABLE_C",
-                    "expectation_name": "range_check",
-                    "expectation_violated": False,
-                },
-            ]
-        }
-        output = capture_reporter_output(TestReporter(), FakeCursor(data))
-        assert output == snapshot
-
-    def test_empty_cursor(self, snapshot):
-        output = capture_reporter_output(TestReporter(), FakeCursor(None))
-        assert output == snapshot
-
-    def test_no_expectations(self, snapshot):
-        data = {"expectations": []}
-        output = capture_reporter_output(TestReporter(), FakeCursor(data))
-        assert output == snapshot
-
-    def test_missing_expectations_key(self, snapshot):
-        data = {"some_other_key": "value"}
-        output = capture_reporter_output(TestReporter(), FakeCursor(data))
-        assert output == snapshot
-
-    def test_missing_expectation_violated_field(self, snapshot):
-        data = {
-            "expectations": [
-                {
-                    "table_name": "DB.SCHEMA.UNKNOWN_STATUS",
-                    "expectation_name": "some_check",
-                }
-            ]
-        }
-        output = capture_reporter_output(TestReporter(), FakeCursor(data))
-        assert output == snapshot
-
-    def test_ansi_codes_in_names(self, snapshot):
-        data = {
-            "expectations": [
-                {
-                    "table_name": "DB.SCHEMA.\x1b[31mRED_TABLE\x1b[0m",
-                    "expectation_name": "\x1b[32mgreen_check\x1b[0m",
-                    "expectation_violated": False,
-                }
-            ]
-        }
-        output = capture_reporter_output(TestReporter(), FakeCursor(data))
-        assert output == snapshot
-
-    def test_non_dict_entries(self, snapshot):
-        data = {
-            "expectations": [
-                "not_a_dict",
-                {
-                    "table_name": "DB.SCHEMA.VALID",
-                    "expectation_name": "valid_check",
-                    "expectation_violated": False,
-                },
-            ]
-        }
-        output = capture_reporter_output(TestReporter(), FakeCursor(data))
-        assert output == snapshot
-
-    def test_process_raises_cli_error_on_failures(self):
-        data = {
-            "expectations": [
-                {
-                    "table_name": "DB.SCHEMA.FAILED",
-                    "expectation_name": "failed_check",
-                    "expectation_violated": True,
-                }
-            ]
-        }
-        reporter = TestReporter()
-        cursor = FakeCursor(data)
-
-        with mock.patch(
-            "snowflake.cli._plugins.dcm.reporters.cli_console.styled_message"
-        ):
-            with pytest.raises(CliError) as exc_info:
-                reporter.process(cursor)
-
-        assert "1 failed" in exc_info.value.message
-
-    def test_process_does_not_raise_on_success(self):
-        data = {
-            "expectations": [
-                {
-                    "table_name": "DB.SCHEMA.PASSED",
-                    "expectation_name": "passed_check",
-                    "expectation_violated": False,
-                }
-            ]
-        }
-        reporter = TestReporter()
-        cursor = FakeCursor(data)
-
-        with mock.patch(
-            "snowflake.cli._plugins.dcm.reporters.cli_console.styled_message"
-        ):
-            reporter.process(cursor)  # Should not raise
-
-
-class TestTestRow:
-    def test_from_dict_with_passing_expectation(self):
-        data = {
-            "table_name": "MY_TABLE",
-            "expectation_name": "my_check",
-            "expectation_violated": False,
-            "expectation_expression": "> 100",
-            "metric_name": "row_count",
-            "value": "500",
-        }
-        row = TestRow.from_dict(data)
-        assert row is not None
-        assert row.table_name == "MY_TABLE"
-        assert row.expectation_name == "my_check"
-        assert row.status == TestStatus.PASS
-        assert row.expectation_expression == "> 100"
-        assert row.metric_name == "row_count"
-        assert row.actual_value == "500"
-
-    def test_from_dict_with_failing_expectation(self):
-        data = {
-            "table_name": "MY_TABLE",
-            "expectation_name": "my_check",
-            "expectation_violated": True,
-        }
-        row = TestRow.from_dict(data)
-        assert row is not None
-        assert row.status == TestStatus.FAIL
-
-    def test_from_dict_with_unknown_status(self):
-        data = {
-            "table_name": "MY_TABLE",
-            "expectation_name": "my_check",
-        }
-        row = TestRow.from_dict(data)
-        assert row is not None
-        assert row.status == TestStatus.UNKNOWN
-
-    def test_from_dict_with_non_dict(self):
-        row = TestRow.from_dict("not a dict")
-        assert row is None
-
-    def test_from_dict_sanitizes_names(self):
-        data = {
-            "table_name": "TABLE\x1b[31mRED\x1b[0m",
-            "expectation_name": "CHECK\x1b[32mGREEN\x1b[0m",
-            "expectation_violated": False,
-        }
-        row = TestRow.from_dict(data)
-        assert row is not None
-        assert "\x1b" not in row.table_name
-        assert "\x1b" not in row.expectation_name
